// src/interfaces/textViewDTO.tsx
export interface barPlotDTO{
    flair: string
    topic: string
    count: number
}

export type lineChartDTO = {
    date: Date
    // flair: string
    "topic-1": number
    "topic-2": number
    "topic-3": number
    "topic-4": number
    "topic-5": number
    "topic-6": number
    "topic-7": number
    "topic-8": number
    "topic-9": number
    "topic-10": number
    "topic-11": number
    "topic-12": number
    "topic-13": number
    "topic-14": number
    "topic-15": number
    "topic-16": number
    "topic-17": number
    "topic-18": number
    "topic-19": number
    "topic-20": number
    "topic-21": number
    "topic-22": number
    "topic-23": number
    "topic-24": number
    "topic-25": number
}